[bug-fix] allow non-default child requirements to override inherited requirements.
This change allows non-default child requirements override parent requirements
for requirements with the same name. Previously, both the parent and
child requirement would be produced, which causes issues during
downstream conda or pip package resolution.

A few other minor changes:
- change the API docs to for Requirements.join to say a warning log
message is produced, where previously it said incorrectly an exception
would be thrown, when there were two requirements with the sane name
- due to the bug fix above, the order of some of the dependencies may
change, for example in the output of Assemble or Tabulate.
- added a test for Spec.compile to verify that default requirements are
applied to a parent environment before a child tries to inherit from it,
and that a child requirement can override the parent requirement for a
requirement with the same name.